ECR-3779 Add no key hashing option to ProofMapIndexProxy
