+++ exonum-java-binding/core/src/main/java/com/exonum/binding/core/storage/indices/ProofMapIndexProxy.java
@@ -44,8 +44,15 @@ import java.util.function.LongSupplier;
  * that a certain key is mapped to a particular value <em>or</em> that there are no mapping for
  * the key in the map.
  *
- * <p>This map is implemented as a Merkle-Patricia tree. It does not permit null keys and values,
- * and requires that keys are 32-byte long.
+ * <p>This map is implemented as a Merkle-Patricia tree. It does not permit null keys and values.
+ *
+ * <p>This map can be instantiated in two ways - either with keys hashing with methods
+ * {@link #newInstance(String, View, Serializer, Serializer)} and
+ * {@link #newInGroupUnsafe(String, byte[], View, Serializer, Serializer)} or with no key hashing
+ * with methods {@link #newInstanceNoKeyHashing(String, View, Serializer, Serializer)} and
+ * {@link #newInGroupUnsafeNoKeyHashing(String, byte[], View, Serializer, Serializer)}. In case of
+ * no key hashing keys are required to be 32-byte long. Note that the former option is considered
+ * a default one.
  *
  * <p>The "destructive" methods of the map, i.e., the one that change the map contents,
  * are specified to throw {@link UnsupportedOperationException} if
@@ -58,7 +65,7 @@ import java.util.function.LongSupplier;
  * <p>When the view goes out of scope, this map is destroyed. Subsequent use of the closed map
  * is prohibited and will result in {@link IllegalStateException}.
  *
- * @param <K> the type of keys in this map. Must be 32-byte long in the serialized form
+ * @param <K> the type of keys in this map
  * @param <V> the type of values in this map
  * @see View
  */
@@ -74,7 +81,7 @@ public final class ProofMapIndexProxy<K, V> extends AbstractIndexProxy implement
    *             [a-zA-Z0-9_]
    * @param view a database view. Must be valid.
    *             If a view is read-only, "destructive" operations are not permitted.
-   * @param keySerializer a serializer of keys, must always produce 32-byte long values
+   * @param keySerializer a serializer of keys
    * @param valueSerializer a serializer of values
    * @param <K> the type of keys in the map
    * @param <V> the type of values in the map
@@ -86,14 +93,68 @@ public final class ProofMapIndexProxy<K, V> extends AbstractIndexProxy implement
       String name, View view, Serializer<K> keySerializer, Serializer<V> valueSerializer) {
     IndexAddress address = IndexAddress.valueOf(name);
     long viewNativeHandle = view.getViewNativeHandle();
-    LongSupplier nativeMapConstructor = () -> nativeCreate(name, viewNativeHandle);
+    LongSupplier nativeMapConstructor = () -> nativeCreate(name, viewNativeHandle, true);
 
-    return getOrCreate(address, view, keySerializer, valueSerializer, nativeMapConstructor);
+    return getOrCreate(address, view, keySerializer, valueSerializer, nativeMapConstructor, true);
+  }
+
+  /**
+   * Creates a ProofMapIndexProxy with no key hashing. Requires that keys are 32-byte long.
+   *
+   * @param name a unique alphanumeric non-empty identifier of this map in the underlying storage:
+   *             [a-zA-Z0-9_]
+   * @param view a database view. Must be valid.
+   *             If a view is read-only, "destructive" operations are not permitted.
+   * @param keySerializer a serializer of keys, must always produce 32-byte long values
+   * @param valueSerializer a serializer of values
+   * @param <K> the type of keys in the map
+   * @param <V> the type of values in the map
+   * @throws IllegalStateException if the view is not valid
+   * @throws IllegalArgumentException if the name is empty
+   * @see StandardSerializers
+   * @see ProofMapIndexProxy
+   */
+  public static <K, V> ProofMapIndexProxy<K, V> newInstanceNoKeyHashing(
+      String name, View view, Serializer<K> keySerializer, Serializer<V> valueSerializer) {
+    IndexAddress address = IndexAddress.valueOf(name);
+    long viewNativeHandle = view.getViewNativeHandle();
+    LongSupplier nativeMapConstructor = () -> nativeCreate(name, viewNativeHandle, false);
+
+    return getOrCreate(address, view, keySerializer, valueSerializer, nativeMapConstructor, false);
   }
 
   /**
    * Creates a new proof map in a <a href="package-summary.html#families">collection group</a>
    * with the given name.
+   *
+   * <p>See a <a href="package-summary.html#families-limitations">caveat</a> on index identifiers.
+   *
+   * @param groupName a name of the collection group
+   * @param mapId an identifier of this collection in the group, see the caveats
+   * @param view a database view
+   * @param keySerializer a serializer of keys
+   * @param valueSerializer a serializer of values
+   * @param <K> the type of keys in the map
+   * @param <V> the type of values in the map
+   * @return a new map proxy
+   * @throws IllegalStateException if the view is not valid
+   * @throws IllegalArgumentException if the name or index id is empty
+   * @see StandardSerializers
+   */
+  public static <K, V> ProofMapIndexProxy<K, V> newInGroupUnsafe(
+      String groupName, byte[] mapId, View view, Serializer<K> keySerializer,
+      Serializer<V> valueSerializer) {
+    IndexAddress address = IndexAddress.valueOf(groupName, mapId);
+    long viewNativeHandle = view.getViewNativeHandle();
+    LongSupplier nativeMapConstructor =
+        () -> nativeCreateInGroup(groupName, mapId, viewNativeHandle, true);
+
+    return getOrCreate(address, view, keySerializer, valueSerializer, nativeMapConstructor, true);
+  }
+
+  /**
+   * Creates a new proof map in a <a href="package-summary.html#families">collection group</a>
+   * with the given name. Requires that keys are 32-byte long.
    *
    * <p>See a <a href="package-summary.html#families-limitations">caveat</a> on index identifiers.
    *
@@ -108,27 +169,26 @@ public final class ProofMapIndexProxy<K, V> extends AbstractIndexProxy implement
    * @throws IllegalStateException if the view is not valid
    * @throws IllegalArgumentException if the name or index id is empty
    * @see StandardSerializers
+   * @see ProofMapIndexProxy
    */
-  public static <K, V> ProofMapIndexProxy<K, V> newInGroupUnsafe(String groupName,
-                                                                 byte[] mapId,
-                                                                 View view,
-                                                                 Serializer<K> keySerializer,
-                                                                 Serializer<V> valueSerializer) {
+  public static <K, V> ProofMapIndexProxy<K, V> newInGroupUnsafeNoKeyHashing(
+      String groupName, byte[] mapId, View view, Serializer<K> keySerializer,
+      Serializer<V> valueSerializer) {
     IndexAddress address = IndexAddress.valueOf(groupName, mapId);
     long viewNativeHandle = view.getViewNativeHandle();
     LongSupplier nativeMapConstructor =
-        () -> nativeCreateInGroup(groupName, mapId, viewNativeHandle);
+        () -> nativeCreateInGroup(groupName, mapId, viewNativeHandle, false);
 
-    return getOrCreate(address, view, keySerializer, valueSerializer, nativeMapConstructor);
+    return getOrCreate(address, view, keySerializer, valueSerializer, nativeMapConstructor, false);
   }
 
   private static <K, V> ProofMapIndexProxy<K, V> getOrCreate(IndexAddress address, View view,
       Serializer<K> keySerializer, Serializer<V> valueSerializer,
-      LongSupplier nativeMapConstructor) {
+      LongSupplier nativeMapConstructor, boolean keyHashing) {
     return view.findOpenIndex(address)
         .map(ProofMapIndexProxy::<K, V>checkCachedInstance)
         .orElseGet(() -> newMapIndexProxy(address, view, keySerializer, valueSerializer,
-            nativeMapConstructor));
+            nativeMapConstructor, keyHashing));
   }
 
   @SuppressWarnings("unchecked") // The compiler is correct: the cache is not type-safe: ECR-3387
@@ -139,9 +199,8 @@ public final class ProofMapIndexProxy<K, V> extends AbstractIndexProxy implement
 
   private static <K, V> ProofMapIndexProxy<K, V> newMapIndexProxy(IndexAddress address, View view,
       Serializer<K> keySerializer, Serializer<V> valueSerializer,
-      LongSupplier nativeMapConstructor) {
-    ProofMapKeyCheckingSerializerDecorator<K> ks =
-        ProofMapKeyCheckingSerializerDecorator.from(keySerializer);
+      LongSupplier nativeMapConstructor, boolean keyHashing) {
+    ProofMapKeyCheckingSerializerDecorator<K> ks = decorateKeySerializer(keySerializer, keyHashing);
     CheckingSerializerDecorator<V> vs = CheckingSerializerDecorator.from(valueSerializer);
 
     NativeHandle mapNativeHandle = createNativeMap(view, nativeMapConstructor);
@@ -149,6 +208,17 @@ public final class ProofMapIndexProxy<K, V> extends AbstractIndexProxy implement
     ProofMapIndexProxy<K, V> map = new ProofMapIndexProxy<>(mapNativeHandle, address, view, ks, vs);
     view.registerIndex(map);
     return map;
+  }
+
+  private static <K> ProofMapKeyCheckingSerializerDecorator<K> decorateKeySerializer(
+      Serializer<K> keySerializer, boolean keyHashing) {
+    if (!keyHashing) {
+      ProofMapKeySizeCheckingSerializerDecorator<K> sizeCheckingSerializerDecorator =
+          ProofMapKeySizeCheckingSerializerDecorator.from(keySerializer);
+      return ProofMapKeyCheckingSerializerDecorator.from(sizeCheckingSerializerDecorator);
+    } else {
+      return ProofMapKeyCheckingSerializerDecorator.from(keySerializer);
+    }
   }
 
   private static NativeHandle createNativeMap(View view, LongSupplier nativeMapConstructor) {
@@ -160,10 +230,10 @@ public final class ProofMapIndexProxy<K, V> extends AbstractIndexProxy implement
     return mapNativeHandle;
   }
 
-  private static native long nativeCreate(String name, long viewNativeHandle);
+  private static native long nativeCreate(String name, long viewNativeHandle, boolean keyHashing);
 
   private static native long nativeCreateInGroup(String groupName, byte[] mapId,
-                                                 long viewNativeHandle);
+                                                 long viewNativeHandle, boolean keyHashing);
 
   private ProofMapIndexProxy(NativeHandle nativeHandle, IndexAddress address, View view,
                              ProofMapKeyCheckingSerializerDecorator<K> keySerializer,
@@ -184,10 +254,11 @@ public final class ProofMapIndexProxy<K, V> extends AbstractIndexProxy implement
   /**
    * {@inheritDoc}
    *
-   * @param key a proof map key, must be 32-byte long when serialized
+   * @param key a proof map key
    * @param value a storage value to associate with the key
    * @throws IllegalStateException if this map is not valid
-   * @throws IllegalArgumentException if the size of the key is not 32 bytes
+   * @throws IllegalArgumentException if the size of the key is not 32 bytes (in case of a
+   *     no key hashing proof map)
    * @throws UnsupportedOperationException if this map is read-only
    */
   @Override
@@ -227,11 +298,11 @@ public final class ProofMapIndexProxy<K, V> extends AbstractIndexProxy implement
    * Returns a proof that there are values mapped to the specified keys or that there are no such
    * mappings.
    *
-   * @param key a proof map key which might be mapped to some value, must be 32-byte long
-   * @param otherKeys other proof map keys which might be mapped to some values, each must be
-   *                  32-byte long
+   * @param key a proof map key which might be mapped to some value
+   * @param otherKeys other proof map keys which might be mapped to some values
    * @throws IllegalStateException if this map is not valid
-   * @throws IllegalArgumentException if the size of any of the keys is not 32 bytes
+   * @throws IllegalArgumentException if the size of any of the keys is not 32 bytes (in case of a
+   *     no key hashing proof map)
    */
   public UncheckedMapProof getProof(K key, K... otherKeys) {
     if (otherKeys.length == 0) {
@@ -246,10 +317,10 @@ public final class ProofMapIndexProxy<K, V> extends AbstractIndexProxy implement
    * Returns a proof that there are values mapped to the specified keys or that there are no such
    * mappings.
    *
-   * @param keys proof map keys which might be mapped to some values, each must be 32-byte long
+   * @param keys proof map keys which might be mapped to some values
    * @throws IllegalStateException if this map is not valid
-   * @throws IllegalArgumentException if the size of any of the keys is not 32 bytes
-   *                                  or keys collection is empty
+   * @throws IllegalArgumentException if the size of any of the keys is not 32 bytes (in case of a
+   *     no key hashing proof map) or keys collection is empty
    */
   public UncheckedMapProof getProof(Collection<? extends K> keys) {
     checkArgument(!keys.isEmpty(), "Keys collection should not be empty");
