+++ exonum-java-binding/core/src/test/java/com/exonum/binding/core/storage/indices/ProofMapIndexProxyIntegrationTest.java
@@ -16,68 +16,17 @@
 
 package com.exonum.binding.core.storage.indices;
 
-import static com.exonum.binding.core.storage.indices.CheckedMapProofMatcher.isValid;
-import static com.exonum.binding.core.storage.indices.MapEntries.putAll;
-import static com.exonum.binding.core.storage.indices.MapTestEntry.absentEntry;
-import static com.exonum.binding.core.storage.indices.MapTestEntry.presentEntry;
-import static com.exonum.binding.core.storage.indices.ProofMapContainsMatcher.provesThatAbsent;
-import static com.exonum.binding.core.storage.indices.ProofMapContainsMatcher.provesThatCorrect;
-import static com.exonum.binding.core.storage.indices.ProofMapContainsMatcher.provesThatPresent;
-import static com.exonum.binding.core.storage.indices.StoragePreconditions.PROOF_MAP_KEY_SIZE;
-import static com.exonum.binding.core.storage.indices.StoragePreconditions.checkProofKey;
-import static com.exonum.binding.core.storage.indices.TestStorageItems.V1;
-import static com.exonum.binding.core.storage.indices.TestStorageItems.V2;
-import static com.exonum.binding.core.storage.indices.TestStorageItems.V3;
-import static com.exonum.binding.core.storage.indices.TestStorageItems.V4;
-import static com.exonum.binding.core.storage.indices.TestStorageItems.values;
-import static com.exonum.binding.test.Bytes.bytes;
-import static com.exonum.binding.test.Bytes.createPrefixed;
-import static com.google.common.base.Preconditions.checkArgument;
-import static java.util.Collections.singletonList;
-import static org.hamcrest.CoreMatchers.containsString;
-import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.CoreMatchers.notNullValue;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.core.IsNot.not;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static java.util.stream.Collectors.toList;
 
-import com.exonum.binding.common.collect.MapEntry;
 import com.exonum.binding.common.hash.HashCode;
-import com.exonum.binding.common.hash.Hashing;
-import com.exonum.binding.common.proofs.map.CheckedMapProof;
-import com.exonum.binding.common.proofs.map.UncheckedMapProof;
 import com.exonum.binding.common.serialization.StandardSerializers;
-import com.exonum.binding.core.proxy.Cleaner;
-import com.exonum.binding.core.proxy.CloseFailuresException;
 import com.exonum.binding.core.storage.database.View;
 import com.exonum.binding.test.Bytes;
-import com.exonum.binding.test.CiOnly;
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.Streams;
-import com.google.common.primitives.UnsignedBytes;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.BitSet;
-import java.util.Collections;
-import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
-import java.util.function.BiConsumer;
-import java.util.function.Consumer;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 import java.util.stream.Stream;
-import org.junit.jupiter.api.Test;
 
 class ProofMapIndexProxyIntegrationTest
-    extends BaseIndexProxyTestable<ProofMapIndexProxy<HashCode, String>> {
-
-  private static final String MAP_NAME = "test_proof_map";
+    extends BaseProofMapIndexProxyIntegrationTest {
 
   private static final List<HashCode> PROOF_KEYS = Stream.of(
       Bytes.bytes(0x00),
@@ -95,717 +44,18 @@ class ProofMapIndexProxyIntegrationTest
       Bytes.bytes(0x10, 0x01),
       Bytes.bytes(0x10, 0x10)
   )
-      .map(ProofMapIndexProxyIntegrationTest::createRawProofKey)
-      .sorted(UnsignedBytes.lexicographicalComparator())
       .map(HashCode::fromBytes)
-      .collect(Collectors.toList());
-
-  static final HashCode PK1 = PROOF_KEYS.get(0);
-  static final HashCode PK2 = PROOF_KEYS.get(1);
-  static final HashCode PK3 = PROOF_KEYS.get(2);
-
-  private static final HashCode INVALID_PROOF_KEY = HashCode.fromString("1234");
-
-  private static final HashCode EMPTY_MAP_INDEX_HASH = HashCode.fromString(
-          "7324b5c72b51bb5d4c180f1109cfd347b60473882145841c39f3e584576296f9");
-
-  @Test
-  void containsKey() {
-    runTestWithView(database::createFork, (map) -> {
-      map.put(PK1, V1);
-      assertTrue(map.containsKey(PK1));
-    });
-  }
-
-  @Test
-  void doesNotContainAbsentKey() {
-    runTestWithView(database::createFork, (map) -> {
-      map.put(PK1, V1);
-      assertFalse(map.containsKey(PK2));
-    });
-  }
-
-  @Test
-  void emptyMapDoesNotContainAbsentKey() {
-    runTestWithView(database::createSnapshot, (map) -> assertFalse(map.containsKey(PK2)));
-  }
-
-  @Test
-  void containsThrowsIfNullKey() {
-    runTestWithView(database::createSnapshot,
-        (map) -> assertThrows(NullPointerException.class, () -> map.containsKey(null)));
-  }
-
-  @Test
-  void containsThrowsIfInvalidKey() {
-    runTestWithView(database::createSnapshot, (map) -> assertThrows(IllegalArgumentException.class,
-        () -> map.containsKey(INVALID_PROOF_KEY)));
-  }
-
-  @Test
-  void putFailsIfSnapshot() {
-    runTestWithView(database::createSnapshot,
-        (map) -> assertThrows(UnsupportedOperationException.class, () -> map.put(PK1, V1)));
-  }
-
-  @Test
-  void putFailsIfInvalidKey() {
-    runTestWithView(database::createFork, (map) -> assertThrows(IllegalArgumentException.class,
-        () -> map.put(INVALID_PROOF_KEY, V1)));
-  }
-
-  @Test
-  void putAllInEmptyMap() {
-    runTestWithView(database::createFork, (map) -> {
-      ImmutableMap<HashCode, String> source = ImmutableMap.of(
-          PK1, V1,
-          PK2, V2
-      );
-
-      map.putAll(source);
-
-      // Check that the map contains all items
-      for (Map.Entry<HashCode, String> entry : source.entrySet()) {
-        HashCode key = entry.getKey();
-        assertTrue(map.containsKey(key));
-        assertThat(map.get(key), equalTo(entry.getValue()));
-      }
-    });
-  }
-
-  @Test
-  void putAllOverwritingEntries() {
-    runTestWithView(database::createFork, (map) -> {
-      map.putAll(ImmutableMap.of(
-          PK1, V1,
-          PK2, V2
-      ));
-
-      ImmutableMap<HashCode, String> replacements = ImmutableMap.of(
-          PK1, V3,
-          PK2, V4
-      );
-
-      map.putAll(replacements);
-
-      // Check that the map contains the recently put entries
-      for (Map.Entry<HashCode, String> entry : replacements.entrySet()) {
-        HashCode key = entry.getKey();
-        assertTrue(map.containsKey(key));
-        assertThat(map.get(key), equalTo(entry.getValue()));
-      }
-    });
-  }
-
-  @Test
-  void get() {
-    runTestWithView(database::createFork, (map) -> {
-      map.put(PK1, V1);
-
-      assertThat(map.get(PK1), equalTo(V1));
-    });
-  }
-
-  @Test
-  void getIndexHash_EmptyMap() {
-    runTestWithView(database::createSnapshot,
-        (map) -> assertThat(map.getIndexHash(), equalTo(EMPTY_MAP_INDEX_HASH)));
-  }
-
-  @Test
-  void getIndexHash_NonEmptyMap() {
-    runTestWithView(database::createFork, (map) -> {
-      map.put(PK1, V1);
-
-      HashCode indexHash = map.getIndexHash();
-      assertThat(indexHash, notNullValue());
-      assertThat(indexHash.bits(), equalTo(Hashing.DEFAULT_HASH_SIZE_BITS));
-      assertThat(indexHash, not(equalTo(EMPTY_MAP_INDEX_HASH)));
-    });
-  }
-
-  @Test
-  void getProof_EmptyMapDoesNotContainSingleKey() {
-    runTestWithView(database::createSnapshot,
-        (map) -> assertThat(map, provesThatAbsent(PK1))
-    );
-  }
-
-  @Test
-  void getProof_SingletonMapContains() {
-    runTestWithView(database::createFork, (map) -> {
-      HashCode key = PK1;
-      String value = V1;
-      map.put(key, value);
-
-      assertThat(map, provesThatPresent(key, value));
-    });
-  }
-
-  @Test
-  void getProof_SingletonMapDoesNotContain() {
-    runTestWithView(database::createFork, (map) -> {
-      map.put(PK1, V1);
-
-      assertThat(map, provesThatAbsent(PK2));
-    });
-  }
-
-  @Test
-  void getProof_FourEntryMap_LastByte_Contains1() {
-    runTestWithView(database::createFork, (map) -> {
-
-      Stream<HashCode> proofKeys = Stream.of(
-          (byte) 0b0000_0000,
-          (byte) 0b0000_0001,
-          (byte) 0b1000_0000,
-          (byte) 0b1000_0001
-      ).map(ProofMapIndexProxyIntegrationTest::createProofKey);
-
-      List<MapEntry<HashCode, String>> entries = createMapEntries(proofKeys);
-
-      putAll(map, entries);
-
-      for (MapEntry<HashCode, String> e : entries) {
-        assertThat(map, provesThatPresent(e.getKey(), e.getValue()));
-      }
-    });
-  }
-
-  @Test
-  void getProof_FourEntryMap_LastByte_Contains2() {
-    runTestWithView(database::createFork, (map) -> {
-      Stream<HashCode> proofKeys = Stream.of(
-          (byte) 0b00,
-          (byte) 0b01,
-          (byte) 0b10,
-          (byte) 0b11
-      ).map(ProofMapIndexProxyIntegrationTest::createProofKey);
-
-      List<MapEntry<HashCode, String>> entries = createMapEntries(proofKeys);
-
-      putAll(map, entries);
-
-      for (MapEntry<HashCode, String> e : entries) {
-        assertThat(map, provesThatPresent(e.getKey(), e.getValue()));
-      }
-    });
-  }
-
-  @Test
-  void getProof_FourEntryMap_FirstByte_Contains() {
-    runTestWithView(database::createFork, (map) -> {
-      byte[] key1 = createRawProofKey();
-      byte[] key2 = createRawProofKey();
-      key2[0] = (byte) 0b0000_0001;
-      byte[] key3 = createRawProofKey();
-      key3[0] = (byte) 0b1000_0000;
-      byte[] key4 = createRawProofKey();
-      key4[0] = (byte) 0b1000_0001;
-
-      List<MapEntry<HashCode, String>> entries = createMapEntries(
-          Stream.of(key1, key2, key3, key4)
-              .map(HashCode::fromBytes)
-      );
-
-      putAll(map, entries);
-
-      for (MapEntry<HashCode, String> e : entries) {
-        assertThat(map, provesThatPresent(e.getKey(), e.getValue()));
-      }
-    });
-  }
-
-  @Test
-  void getProof_FourEntryMap_FirstAndLastByte_Contains() {
-    runTestWithView(database::createFork, (map) -> {
-      byte[] key1 = createRawProofKey();  // 000…0
-      byte[] key2 = createRawProofKey();  // 100…0
-      key2[0] = (byte) 0x01;
-      byte[] key3 = createRawProofKey((byte) 0x80);  // 000…01
-      byte[] key4 = createRawProofKey((byte) 0x80);  // 100…01
-      key4[0] = (byte) 0x01;
-
-      List<MapEntry<HashCode, String>> entries = createMapEntries(
-          Stream.of(key1, key2, key3, key4)
-              .map(HashCode::fromBytes)
-      );
-
-      putAll(map, entries);
-
-      for (MapEntry<HashCode, String> e : entries) {
-        assertThat(map, provesThatPresent(e.getKey(), e.getValue()));
-      }
-    });
-  }
-
-  @Test
-  void getProof_MultiEntryMapContains() {
-    runTestWithView(database::createFork, (map) -> {
-      List<MapEntry<HashCode, String>> entries = createSortedMapEntries();
-      putAll(map, entries);
-
-      for (MapEntry<HashCode, String> e : entries) {
-        assertThat(map, provesThatPresent(e.getKey(), e.getValue()));
-      }
-    });
-  }
-
-  @Test
-  void getProof_MultiEntryMapDoesNotContain() {
-    runTestWithView(database::createFork, (map) -> {
-      List<MapEntry<HashCode, String>> entries = createSortedMapEntries();
-      putAll(map, entries);
-
-      byte[] allOnes = new byte[PROOF_MAP_KEY_SIZE];
-      Arrays.fill(allOnes, UnsignedBytes.checkedCast(0xFF));
-
-      List<HashCode> otherKeys = ImmutableList.of(
-          HashCode.fromBytes(allOnes),  // [11…1]
-          createProofKey("PK1001"),
-          createProofKey("PK1002"),
-          createProofKey("PK100500")
-      );
-
-      for (HashCode key : otherKeys) {
-        assertThat(map, provesThatAbsent(key));
-      }
-    });
-  }
-
-  @CiOnly
-  @Test
-  /*
-    Takes quite a lot of time (validating 257 proofs), but it's an integration test, isn't it? :-)
-    Consider adding a similar test for left-leaning MPT
-   */
-  void getProof_MapContainsRightLeaningMaxHeightMpt() {
-    runTestWithView(database::createFork, (map) -> {
-      List<MapEntry<HashCode, String>> entries = createEntriesForRightLeaningMpt();
-      putAll(map, entries);
-
-      for (MapEntry<HashCode, String> e : entries) {
-        assertThat(map, provesThatPresent(e.getKey(), e.getValue()));
-      }
-    });
-  }
-
-  @Test
-  void getMultiProof_EmptyMapDoesNotContainSeveralKeys() {
-    runTestWithView(database::createSnapshot, (map) ->
-        assertThat(map, provesThatAbsent(PK1, PK2)));
-  }
-
-  @Test
-  void getMultiProof_SingletonMapDoesNotContainSeveralKeys() {
-    runTestWithView(database::createFork, (map) -> {
-      map.put(PK1, V1);
-
-      assertThat(map, provesThatAbsent(PK2, PK3));
-    });
-  }
-
-  @Test
-  void getMultiProof_SingletonMapBothContainsAndDoesNot() {
-    runTestWithView(database::createFork, (map) -> {
-      ImmutableMap<HashCode, String> source = ImmutableMap.of(
-          PK1, V1
-      );
-
-      map.putAll(source);
-
-      assertThat(map, provesThatCorrect(presentEntry(PK1, V1), absentEntry(PK2)));
-    });
-  }
-
-  @Test
-  void getMultiProof_TwoElementMapContains() {
-    runTestWithView(database::createFork, (map) -> {
-      ImmutableMap<HashCode, String> source = ImmutableMap.of(
-          PK1, V1,
-          PK2, V2
-      );
-
-      map.putAll(source);
-
-      assertThat(map, provesThatCorrect(presentEntry(PK1, V1), presentEntry(PK2, V2)));
-    });
-  }
-
-  @Test
-  void getMultiProof_FourEntryMap_LastByte_Contains1() {
-    runTestWithView(database::createFork, (map) -> {
-
-      Stream<HashCode> proofKeys = Stream.of(
-          (byte) 0b0000_0000,
-          (byte) 0b0000_0001,
-          (byte) 0b1000_0000,
-          (byte) 0b1000_0001
-      ).map(ProofMapIndexProxyIntegrationTest::createProofKey);
-
-      List<MapEntry<HashCode, String>> entries = createMapEntries(proofKeys);
-
-      putAll(map, entries);
-
-      assertThat(map, provesThatPresent(entries));
-    });
-  }
-
-  @Test
-  void getMultiProof_FourEntryMap_LastByte_Contains2() {
-    runTestWithView(database::createFork, (map) -> {
-      Stream<HashCode> proofKeys = Stream.of(
-          (byte) 0b00,
-          (byte) 0b01,
-          (byte) 0b10,
-          (byte) 0b11
-      ).map(ProofMapIndexProxyIntegrationTest::createProofKey);
-
-      List<MapEntry<HashCode, String>> entries = createMapEntries(proofKeys);
-
-      putAll(map, entries);
-
-      assertThat(map, provesThatPresent(entries));
-    });
-  }
-
-  @Test
-  void getMultiProof_FourEntryMap_FirstByte_Contains() {
-    runTestWithView(database::createFork, (map) -> {
-      byte[] key1 = createRawProofKey();
-      byte[] key2 = createRawProofKey();
-      key2[0] = (byte) 0b0000_0001;
-      byte[] key3 = createRawProofKey();
-      key3[0] = (byte) 0b1000_0000;
-      byte[] key4 = createRawProofKey();
-      key4[0] = (byte) 0b1000_0001;
-
-      List<MapEntry<HashCode, String>> entries = createMapEntries(
-          Stream.of(key1, key2, key3, key4)
-              .map(HashCode::fromBytes)
-      );
-
-      putAll(map, entries);
-
-      assertThat(map, provesThatPresent(entries));
-    });
-  }
-
-  @Test
-  void getMultiProof_FourEntryMap_FirstAndLastByte_Contains() {
-    runTestWithView(database::createFork, (map) -> {
-      byte[] key1 = createRawProofKey();  // 000…0
-      byte[] key2 = createRawProofKey();  // 100…0
-      key2[0] = (byte) 0x01;
-      byte[] key3 = createRawProofKey((byte) 0x80);  // 000…01
-      byte[] key4 = createRawProofKey((byte) 0x80);  // 100…01
-      key4[0] = (byte) 0x01;
-
-      List<MapEntry<HashCode, String>> entries = createMapEntries(
-          Stream.of(key1, key2, key3, key4)
-              .map(HashCode::fromBytes)
-      );
-
-      putAll(map, entries);
-
-      assertThat(map, provesThatPresent(entries));
-    });
-  }
-
-  @Test
-  void getMultiProof_SortedMultiEntryMapContains() {
-    runTestWithView(database::createFork, (map) -> {
-      List<MapEntry<HashCode, String>> entries = createSortedMapEntries();
-      putAll(map, entries);
-
-      assertThat(map, provesThatPresent(entries));
-    });
-  }
-
-  @Test
-  void getMultiProof_FourEntryMap_DoesNotContain() {
-    runTestWithView(database::createFork, (map) -> {
-      /*
-       This map will have the following structure:
-                   <00xxxx>
-                   /        \
-           <00|00xx>          <00|10xx>
-          /         \        /         \
-       <0000|01>  <0000|11>  <0010|00>   <0010|10>
-      */
-      List<MapEntry<HashCode, String>> entries = createMapEntries(
-          Stream.of(
-              proofKeyFromPrefix("0000|01"),
-              proofKeyFromPrefix("0000|11"),
-              proofKeyFromPrefix("0010|00"),
-              proofKeyFromPrefix("0010|10")
-          )
-      );
-
-      putAll(map, entries);
-
-      List<HashCode> proofKeys = Arrays.asList(
-          // Should be rejected on root level
-          proofKeyFromPrefix("01|0000"),
-          // Should be rejected on intermediate level
-          proofKeyFromPrefix("00|01"),
-          proofKeyFromPrefix("00|11"),
-          // Should be rejected on leaf level
-          proofKeyFromPrefix("0000|00"),
-          proofKeyFromPrefix("0000|10"),
-          proofKeyFromPrefix("0010|01"),
-          proofKeyFromPrefix("0010|11")
-      );
-
-      assertThat(map, provesThatAbsent(proofKeys));
-    });
-  }
-
-  @Test
-  void remove() {
-    runTestWithView(database::createFork, (map) -> {
-      map.put(PK1, V1);
-      map.remove(PK1);
-
-      assertNull(map.get(PK1));
-      assertFalse(map.containsKey(PK1));
-    });
-  }
-
-  @Test
-  void removeFailsIfSnapshot() {
-    runTestWithView(database::createSnapshot,
-        (map) -> assertThrows(UnsupportedOperationException.class, () -> map.remove(PK1)));
-  }
-
-  @Test
-  void removeFailsIfInvalidKey() {
-    runTestWithView(database::createFork,
-        (map) -> assertThrows(IllegalArgumentException.class, () -> map.remove(INVALID_PROOF_KEY)));
-  }
-
-  @Test
-  void keysTest() {
-    runTestWithView(database::createFork, (map) -> {
-      List<MapEntry<HashCode, String>> entries = createSortedMapEntries();
-
-      putAll(map, entries);
-
-      Iterator<HashCode> keysIterator = map.keys();
-      List<HashCode> keysFromIter = ImmutableList.copyOf(keysIterator);
-      List<HashCode> keysInMap = MapEntries.extractKeys(entries);
-
-      // Keys must appear in a lexicographical order.
-      assertThat(keysFromIter, equalTo(keysInMap));
-    });
-  }
-
-  @Test
-  void valuesTest() {
-    runTestWithView(database::createFork, (map) -> {
-      List<MapEntry<HashCode, String>> entries = createSortedMapEntries();
-
-      putAll(map, entries);
-
-      Iterator<String> valuesIterator = map.values();
-      List<String> valuesFromIter = ImmutableList.copyOf(valuesIterator);
-      List<String> valuesInMap = MapEntries.extractValues(entries);
-
-      // Values must appear in a lexicographical order of keys.
-      assertThat(valuesFromIter, equalTo(valuesInMap));
-    });
-  }
-
-  @Test
-  void entriesTest() {
-    runTestWithView(database::createFork, (map) -> {
-      List<MapEntry<HashCode, String>> entries = createSortedMapEntries();
-
-      putAll(map, entries);
-
-      Iterator<MapEntry<HashCode, String>> entriesIterator = map.entries();
-      List<MapEntry> entriesFromIter = ImmutableList.copyOf(entriesIterator);
-      // Entries must appear in a lexicographical order of keys.
-      assertThat(entriesFromIter, equalTo(entries));
-    });
-  }
-
-  @Test
-  void clearEmptyHasNoEffect() {
-    runTestWithView(database::createFork, ProofMapIndexProxy::clear);
-  }
-
-  @Test
-  void clearNonEmptyRemovesAllValues() {
-    runTestWithView(database::createFork, (map) -> {
-      List<MapEntry<HashCode, String>> entries = createSortedMapEntries();
-
-      putAll(map, entries);
-
-      map.clear();
-
-      for (MapEntry<HashCode, ?> e : entries) {
-        assertFalse(map.containsKey(e.getKey()));
-      }
-    });
-  }
-
-  @Test
-  void clearFailsIfSnapshot() {
-    runTestWithView(database::createSnapshot, (map) -> {
-      assertThrows(UnsupportedOperationException.class, () -> map.clear());
-    });
-  }
-
-  @Test
-  void isEmptyShouldReturnTrueForEmptyMap() {
-    runTestWithView(database::createSnapshot, (map) -> assertTrue(map.isEmpty()));
-  }
-
-  @Test
-  void isEmptyShouldReturnFalseForNonEmptyMap() {
-    runTestWithView(database::createFork, (map) -> {
-      map.put(PK1, V1);
-
-      assertFalse(map.isEmpty());
-    });
-  }
-
-  @Test
-  void getProofFromSingleKey() {
-    runTestWithView(database::createFork, (map) -> {
-      map.put(PK1, V1);
-
-      UncheckedMapProof proof = map.getProof(PK1);
-      CheckedMapProof checkedProof = proof.check();
-
-      assertThat(checkedProof, isValid(singletonList(presentEntry(PK1, V1))));
-    });
-  }
-
-  @Test
-  void getProofFromVarargs() {
-    runTestWithView(database::createFork, (map) -> {
-      map.put(PK1, V1);
-      map.put(PK2, V2);
-
-      UncheckedMapProof proof = map.getProof(PK1, PK2);
-      CheckedMapProof checkedProof = proof.check();
-
-      assertThat(
-          checkedProof, isValid(Arrays.asList(presentEntry(PK1, V1), presentEntry(PK2, V2))));
-    });
-  }
-
-  @Test
-  void getProofFromEmptyCollection() {
-    runTestWithView(database::createFork, (map) -> {
-      map.put(PK1, V1);
-
-      IllegalArgumentException thrown = assertThrows(IllegalArgumentException.class, () -> {
-        UncheckedMapProof proof = map.getProof(Collections.emptyList());
-      });
-      assertThat(thrown.getLocalizedMessage(),
-          containsString("Keys collection should not be empty"));
-    });
-  }
-
-  @Test
-  void getProofFromCollection() {
-    runTestWithView(database::createFork, (map) -> {
-      map.put(PK1, V1);
-
-      UncheckedMapProof proof = map.getProof(singletonList(PK1));
-      CheckedMapProof checkedProof = proof.check();
-
-      assertThat(checkedProof, isValid(singletonList(presentEntry(PK1, V1))));
-    });
-  }
-
-  /**
-   * Returns a new key with the given prefix.
-   *
-   * @param prefix a key prefix — from the least significant bit to the most significant,
-   *        i.e., "00 01" is 8, "10 00" is 1.
-   *        May contain spaces, underscores or bars (e.g., "00 01|01 11" and "11_10"
-   *        are valid strings).
-   */
-  private static HashCode proofKeyFromPrefix(String prefix) {
-    prefix = filterBitPrefix(prefix);
-    byte[] key = keyFromString(prefix);
-    return HashCode.fromBytes(key);
-  }
-
-  /**
-   * Replaces spaces that may be used to separate groups of binary digits.
-   */
-  private static String filterBitPrefix(String prefix) {
-    String filtered = prefix.replaceAll("[ _|]", "");
-    // Check that the string is correct
-    assert filtered.matches("[01]*");
-    assert filtered.length() <= PROOF_MAP_KEY_SIZE;
-    return filtered;
-  }
-
-  /**
-   * Creates a 32-byte key from the bit prefix.
-   */
-  private static byte[] keyFromString(String prefix) {
-    BitSet keyPrefixBits = new BitSet(prefix.length());
-    for (int i = 0; i < prefix.length(); i++) {
-      char bit = prefix.charAt(i);
-      if (bit == '1') {
-        keyPrefixBits.set(i);
-      }
-    }
-    return createPrefixed(keyPrefixBits.toByteArray(), PROOF_MAP_KEY_SIZE);
-  }
-
-  /**
-   * Create a proof key of length 32 with the specified suffix.
-   *
-   * @param suffix a key suffix. Must be shorter than or equal to 32 bytes in UTF-8.
-   * @return a key, starting with zeroes and followed by the specified suffix encoded in UTF-8
-   */
-  private static HashCode createProofKey(String suffix) {
-    byte[] suffixBytes = bytes(suffix);
-    return createProofKey(suffixBytes);
-  }
-
-  private static HashCode createProofKey(byte... suffixBytes) {
-    byte[] proofKey = createRawProofKey(suffixBytes);
-    return HashCode.fromBytes(proofKey);
-  }
-
-  private static byte[] createRawProofKey(byte... suffixBytes) {
-    checkArgument(suffixBytes.length <= PROOF_MAP_KEY_SIZE);
-    byte[] proofKey = new byte[PROOF_MAP_KEY_SIZE];
-    System.arraycopy(suffixBytes, 0, proofKey, PROOF_MAP_KEY_SIZE - suffixBytes.length,
-        suffixBytes.length);
-    return checkProofKey(proofKey);
-  }
-
-  private static void runTestWithView(Function<Cleaner, View> viewFactory,
-      Consumer<ProofMapIndexProxy<HashCode, String>> mapTest) {
-    runTestWithView(viewFactory, (ignoredView, map) -> mapTest.accept(map));
-  }
-
-  private static void runTestWithView(
-      Function<Cleaner, View> viewFactory,
-      BiConsumer<View, ProofMapIndexProxy<HashCode, String>> mapTest) {
-    try (Cleaner cleaner = new Cleaner()) {
-      View view = viewFactory.apply(cleaner);
-      ProofMapIndexProxy<HashCode, String> map = createProofMap(MAP_NAME, view);
-
-      mapTest.accept(view, map);
-    } catch (CloseFailuresException e) {
-      throw new AssertionError("Unexpected exception", e);
-    }
+      .collect(toList());
+
+  @Override
+  List<HashCode> getProofKeys() {
+    return PROOF_KEYS;
   }
 
   @Override
   ProofMapIndexProxy<HashCode, String> create(String name, View view) {
-    return createProofMap(name, view);
+    return ProofMapIndexProxy.newInstance(name, view, StandardSerializers.hash(),
+        StandardSerializers.string());
   }
 
   @Override
@@ -813,98 +63,5 @@ class ProofMapIndexProxyIntegrationTest
       View view) {
     return ProofMapIndexProxy.newInGroupUnsafe(groupName, idInGroup, view,
         StandardSerializers.hash(), StandardSerializers.string());
-  }
-
-  @Override
-  StorageIndex createOfOtherType(String name, View view) {
-    return ListIndexProxy.newInstance(name, view, StandardSerializers.string());
-  }
-
-  @Override
-  Object getAnyElement(ProofMapIndexProxy<HashCode, String> index) {
-    return index.get(PK1);
-  }
-
-  @Override
-  void update(ProofMapIndexProxy<HashCode, String> index) {
-    index.put(PK1, V1);
-  }
-
-  private static ProofMapIndexProxy<HashCode, String> createProofMap(String name, View view) {
-    return ProofMapIndexProxy.newInstance(name, view, StandardSerializers.hash(),
-        StandardSerializers.string());
-  }
-
-  /**
-   * Creates `numOfEntries` map entries, sorted by key:
-   * [(00…0PK1, V1), (00…0PK2, V2), … (00…0PKi, Vi)].
-   */
-  private List<MapEntry<HashCode, String>> createSortedMapEntries() {
-    // Use PROOF_KEYS which are already sorted.
-    return createMapEntries(PROOF_KEYS.stream());
-  }
-
-  /**
-   * Creates map entries with the given keys. Uses values
-   * from {@linkplain TestStorageItems#values} in a round-robin fashion.
-   */
-  private List<MapEntry<HashCode, String>> createMapEntries(Stream<HashCode> proofKeys) {
-    Stream<HashCode> keys = proofKeys.distinct();
-    Stream<String> roundRobinValues = IntStream.range(0, Integer.MAX_VALUE)
-        .mapToObj(i -> values.get(i % values.size()));
-    return Streams.zip(keys, roundRobinValues, MapEntry::valueOf)
-        .collect(Collectors.toList());
-  }
-
-  /**
-   * Creates 257 entries for a ProofMap that, when added to it, will make the underlying
-   * Merkle-Patricia tree of the maximum height (256). Leaf nodes will be at depths
-   * ranging from 1 to 256.
-   *
-   * Bits of 32-byte keys:
-   * 00…0000
-   * 100…000
-   * 0100…00
-   * 00100…0
-   * …
-   * 00…0100
-   * 00…0010
-   * 00…0001.
-   *
-   * When all the keys above are added to the ProofMap, the underlying Merkle-Patricia tree
-   * has the following structure (only key bits in leaf nodes are shown; the intermediate
-   * nodes are shown as 'o' character):
-   *
-   *                    o — the root node
-   *                   / \
-   *                  o   100…000 — a leaf node
-   *                 / \
-   *                o   0100…00
-   *               / \
-   *              o   00100…0
-   *             / \
-   *            …   00010…0
-   *           /
-   *          o — an intermediate node with key prefix 00…0 of size 255 bits.
-   *         / \
-   *  00…0000  00…0001 — leaf nodes at depth 256 with a common prefix of 255 bits.
-   */
-  private static List<MapEntry<HashCode, String>> createEntriesForRightLeaningMpt() {
-    int numKeyBits = Byte.SIZE * PROOF_MAP_KEY_SIZE;
-    BitSet keyBits = new BitSet(numKeyBits);
-    int numEntries = numKeyBits + 1;
-    List<MapEntry<HashCode, String>> entries = new ArrayList<>(numEntries);
-    entries.add(MapEntry.valueOf(HashCode.fromBytes(new byte[PROOF_MAP_KEY_SIZE]), V1));
-
-    for (int i = 0; i < numKeyBits; i++) {
-      keyBits.set(i);
-      byte[] key = createPrefixed(keyBits.toByteArray(), PROOF_MAP_KEY_SIZE);
-      String value = values.get(i % values.size());
-      entries.add(MapEntry.valueOf(HashCode.fromBytes(key), value));
-      keyBits.clear(i);
-      assert keyBits.length() == 0;
-    }
-
-    return entries;
   }
 }
